+++ src/module/layer/GaodeLayer/layerSource.ts
@@ -8,15 +8,15 @@ export const GaodeLayerTypeUrls: Record<GaodeLayerTypeEnum, Array<string>> = {
     'https://webrd04.is.autonavi.com/appmaptile?lang=zh_cn&size=1&scale=1&style=8&x={x}&y={y}&z={z}'
   ],
   img: [
-    'http://webst01.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}',
-    'http://webst02.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}',
-    'http://webst03.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}',
-    'http://webst04.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}'
+    'https://webst01.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}',
+    'https://webst02.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}',
+    'https://webst03.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}',
+    'https://webst04.is.autonavi.com/appmaptile?style=6&x={x}&y={y}&z={z}'
   ],
   road: [
-    'http://webst01.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8',
-    'http://webst02.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8',
-    'http://webst03.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8',
-    'http://webst04.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8'
+    'https://webst01.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8',
+    'https://webst02.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8',
+    'https://webst03.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8',
+    'https://webst04.is.autonavi.com/appmaptile?x={x}&y={y}&z={z}&lang=zh_cn&size=1&scale=1&style=8'
   ]
 }